chore(DataStore): migrate functionality over to CascadeDeleteOperation
